SONAR-4736 fix log when converted more than 21M violations (status in percents was negative)

+++ sonar-server/src/main/java/org/sonar/server/db/migrations/violation/Progress.java
@@ -24,6 +24,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.TimerTask;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * This task logs every minute the status of migration. It is destroyed
@@ -34,18 +35,18 @@ class Progress extends TimerTask {
   static final String THREAD_NAME = "Violation Migration Progress";
   static final long DELAY_MS = 60000L;
 
-  private final AtomicInteger counter = new AtomicInteger(0);
+  private final AtomicLong counter = new AtomicLong(0L);
   private final Logger logger;
-  private final int totalViolations;
+  private final long totalViolations;
   private final long start;
 
-  Progress(int totalViolations, Logger logger, long startDate) {
+  Progress(long totalViolations, Logger logger, long startDate) {
     this.totalViolations = totalViolations;
     this.logger = logger;
     this.start = startDate;
   }
 
-  Progress(int totalViolations) {
+  Progress(long totalViolations) {
     this(totalViolations, LoggerFactory.getLogger(Progress.class), System.currentTimeMillis());
   }
 
@@ -55,12 +56,12 @@ class Progress extends TimerTask {
 
   @Override
   public void run() {
-    int totalIssues = counter.get();
+    long totalIssues = counter.get();
     long durationMinutes = (System.currentTimeMillis() - start) / 60000L;
-    int percents = (100 * totalIssues) / totalViolations;
+    int percents = (int)((100L * totalIssues) / totalViolations);
     if (totalIssues>0 && durationMinutes > 0) {
       int frequency = (int) (totalIssues / durationMinutes);
-      int remaining = (totalViolations - totalIssues) / frequency;
+      long remaining = (totalViolations - totalIssues) / frequency;
       logger.info(String.format(
         "%d%% [%d/%d violations, %d minutes remaining]", percents, totalIssues, totalViolations, remaining)
       );
